make erc20 cache warmup phase less aggressive
As the list of erc20 tokens grew, the cache warmup phase took longer
and longer and caused the request that triggered the warmup to timeout
and put a lot of stress on the service. this makes the initial request
to trigger the cache warmup phase return nothing and not wait for the
warmup to finish, and the warmup will send TokenPayment push
notifications as it finds new tokens to trigger the clients to refresh
their token data.